backend/requirements.txt
# AWS SDK for Nova Sonic Bidirectional Streaming
boto3>=1.35.0
botocore>=1.35.0

# FastAPI for WebSocket server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0

# Environment variables
python-dotenv>=1.0.0

# CORS support (included with FastAPI/Starlette)
starlette>=0.40.0
